lantiq: add support for AVM FRITZ!Box 3490
The FRITZ!Box 3490 actually contains two SoCs, one Lantiq with no WiFi and
one QCA558 with 5GHz and 2.4GHz WiFi (via QCA9880). Only the Lantiq has
access to the flash memory, the QCA runs fully from RAM.

Specifications
--------------

  - Lantiq 500 MHz
  - 256MiB RAM
  - 512MiB NAND
  - 8MBit Flash
  - QCA9558 720MHz and 2.4GHz WiFi
  - QCA9880 (5GHz only)
  - 128MiB RAM

Remarks
-------

Some kernel patches are required to add support for the QCA SoC.

Installation
------------

Use the eva_ramboot.py script to boot the initramfs image. Then, transfer
the sysupgrade image to the device and run sysupgrade to flash it to the
NAND.

Signed-off-by: Andreas Böhler <dev@aboehler.at>