ath79: add support for the offloading CPU in AVM FRITZ!Box 3490
The AVM FRITZ!Box 3490 is a device with two SoC: Lantiq and ath79.
While the Lantiq target has access to flash and USB, the ath79 target
only has Ethernet access. Thus, for the FRITZ!Box 3490 only an initramfs
image is required that is uploaded to the ath79 by a firmware uploader
on the ath79 target. A stage 1 firmware is uploaded via MDIO which
accepts a stage 2 firmware via Ethernet. This is the stage 2 firmware.

In order to configure the ath79 instance, a separate config downloader
takes care of EEPROM files and configuration data.